Split Extensions into two classes
