+++ src/xunit.analyzers/SyntaxExtensions.cs
@@ -1,13 +1,11 @@
-﻿using System.Collections.Immutable;
-using System.Linq;
-using System.Threading;
+﻿using System.Threading;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace Xunit.Analyzers
 {
-    static class Extensions
+    internal static class SyntaxExtensions
     {
         internal static bool ContainsAttributeType(this SyntaxList<AttributeListSyntax> attributeLists, SemanticModel semanticModel, INamedTypeSymbol attributeType, bool exactMatch = false)
         {
@@ -23,22 +21,6 @@ namespace Xunit.Analyzers
             return false;
         }
 
-        internal static bool ContainsAttributeType(this ImmutableArray<AttributeData> attributes, INamedTypeSymbol attributeType, bool exactMatch = false)
-        {
-            return attributes.Any(a => attributeType.IsAssignableFrom(a.AttributeClass, exactMatch));
-        }
-
-        internal static ImmutableArray<ISymbol> GetInheritedAndOwnMembers(this ITypeSymbol symbol, string name = null)
-        {
-            var builder = ImmutableArray.CreateBuilder<ISymbol>();
-            while (symbol != null)
-            {
-                builder.AddRange(name == null ? symbol.GetMembers() : symbol.GetMembers(name));
-                symbol = symbol.BaseType;
-            }
-            return builder.ToImmutable();
-        }
-
         internal static SimpleNameSyntax GetSimpleName(this InvocationExpressionSyntax invocation)
         {
             switch (invocation.Expression)
@@ -52,26 +34,13 @@ namespace Xunit.Analyzers
             return null;
         }
 
-        internal static bool IsAssignableFrom(this ITypeSymbol targetType, ITypeSymbol sourceType, bool exactMatch = false)
+        internal static bool IsEnumValueExpression(this ExpressionSyntax expression, SemanticModel semanticModel, CancellationToken cancellationToken = default(CancellationToken))
         {
-            if (targetType != null)
-            {
-                while (sourceType != null)
-                {
-                    if (sourceType.Equals(targetType))
-                        return true;
+            if (!expression.IsKind(SyntaxKind.SimpleMemberAccessExpression))
+                return false;
 
-                    if (exactMatch)
-                        return false;
-
-                    if (targetType.TypeKind == TypeKind.Interface)
-                        return sourceType.AllInterfaces.Any(i => i.Equals(targetType));
-
-                    sourceType = sourceType.BaseType;
-                }
-            }
-
-            return false;
+            var symbol = semanticModel.GetSymbolInfo(expression, cancellationToken).Symbol;
+            return symbol?.Kind == SymbolKind.Field && symbol.ContainingType.TypeKind == TypeKind.Enum;
         }
 
         internal static bool IsNameofExpression(this ExpressionSyntax expression, SemanticModel semanticModel, CancellationToken cancellationToken = default(CancellationToken))
@@ -89,15 +58,6 @@ namespace Xunit.Analyzers
             // A real nameof expression doesn't have a matching symbol, but it does have the string type
             return semanticModel.GetSymbolInfo(expression, cancellationToken).Symbol == null &&
                 semanticModel.GetTypeInfo(expression, cancellationToken).Type?.SpecialType == SpecialType.System_String;
-        }
-
-        internal static bool IsEnumValueExpression(this ExpressionSyntax expression, SemanticModel semanticModel, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            if (!expression.IsKind(SyntaxKind.SimpleMemberAccessExpression))
-                return false;
-
-            var symbol = semanticModel.GetSymbolInfo(expression, cancellationToken).Symbol;
-            return symbol?.Kind == SymbolKind.Field && symbol.ContainingType.TypeKind == TypeKind.Enum;
         }
     }
 }
